Adding UpdatedTimestamp and CreatedTimestamp to mixin them separated
