+++ lib/mongoid.rb
@@ -80,6 +80,8 @@ require "mongoid/safety"
 require "mongoid/scope"
 require "mongoid/serialization"
 require "mongoid/state"
+require "mongoid/created_timestamp"
+require "mongoid/updated_timestamp"
 require "mongoid/timestamps"
 require "mongoid/validations"
 require "mongoid/versioning"
